change type and fix ts error
