+++ packages/next-admin/src/components/Cell.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { ReactNode } from "react";
 
 import { ListDataFieldValue } from "../types";
 import Link from "next/link";
@@ -6,16 +6,19 @@ import clsx from "clsx";
 import { useConfig } from "../context/ConfigContext";
 
 type Props = {
-  cell: ListDataFieldValue;
+  cell: ListDataFieldValue | ReactNode;
 };
 
 export default function Cell({ cell }: Props) {
   const { basePath } = useConfig()
 
-  if (cell !== null) {
+  const isReactNode = (cell: ListDataFieldValue | ReactNode): cell is ReactNode => {
+    return React.isValidElement(cell)
+  }
+  if (cell && cell !== null) {
     if (React.isValidElement(cell)) {
       return cell;
-    } else if (typeof cell === "object") {
+    } else if (typeof cell === "object" && !isReactNode(cell)) {
       if (cell.type === "link") {
         return (
           <Link
@@ -56,10 +59,8 @@ export default function Cell({ cell }: Props) {
         <div
           className={clsx(
             "inline-flex items-center rounded-md px-2 py-1 text-xs font-medium",
-            {
-              "bg-indigo-50 text-indigo-500": cell === true,
-              "bg-neutral-50 text-neutral-600": cell === false,
-            }
+            cell ? "bg-indigo-50 text-indigo-500" :
+              "bg-neutral-50 text-neutral-600"
           )}
         >
           <p>{cell.toString()}</p>
