+++ packages/next-admin/src/components/NextAdmin.tsx
@@ -1,10 +1,7 @@
 import Head from "next/head";
 import Link from "next/link";
 import NextNProgress from 'nextjs-progressbar';
-import { ReactNode } from "react";
-
-import { Prisma } from "@prisma/client";
-import { Field, ListData, ListDataItem, Model, ModelName, Schema } from "../types";
+import { AdminComponentProps, CustomUIProps } from "../types";
 import { getSchemaForResource } from "../utils/jsonSchema";
 import Dashboard from "./Dashboard";
 import Form from "./Form";
@@ -12,45 +9,6 @@ import List from "./List";
 import Menu from "./Menu";
 import Message from "./Message";
 import { ConfigProvider } from "../context/ConfigContext";
-import { PropertyValidationError } from "../exceptions/ValidationError";
-
-export type ListComponentFieldsOptions<T extends ModelName> = {
-  [P in Field<T>]?: {
-    formatter?: (item: ListDataItem<ModelName>) => ReactNode;
-  };
-};
-
-export type AdminComponentOptions<T extends ModelName> = {
-  model?: {
-    [P in T]?: {
-      toString?: (item: Model<P>[number]) => string;
-      list?: {
-        fields: ListComponentFieldsOptions<P>;
-      };
-    };
-  };
-};
-
-export type AdminComponentProps = {
-  basePath: string;
-  schema: Schema;
-  data?: ListData<ModelName>;
-  resource: ModelName;
-  message?: {
-    type: "success" | "info";
-    content: string;
-  };
-  error?: string;
-  validation?: PropertyValidationError[];
-  resources?: ModelName[];
-  total?: number;
-  dmmfSchema: Prisma.DMMF.Field[];
-  options?: AdminComponentOptions<ModelName>;
-};
-
-export type CustomUIProps = {
-  dashboard?: JSX.Element | (() => JSX.Element);
-};
 
 // Components
 export function NextAdmin({
