+++ packages/next-admin/src/tests/serverUtils.test.ts
@@ -25,7 +25,7 @@ describe("fillRelationInSchema", () => {
     ]);
     // @ts-expect-error
     const result = await fillRelationInSchema(schema, prismaMock, "Post", {});
-    expect(result.definitions.Post.properties.authorId.enum).toEqual([
+    expect(result.definitions.Post.properties.authorId?.enum).toEqual([
       { label: 1, value: 1 },
       { label: 2, value: 2 },
     ]);
